refactor(canvas): move chat settings into settings domain
