--- apps/canvas-workspace/src/renderer/src/components/chat/ModelSettings.tsx
@@ -1,3 +1,0 @@
-import './ModelSettings.css';
-
-export { ModelsSection } from './ModelsSection';
