+++ apps/canvas-workspace/src/renderer/src/components/settings/ModelSettings/ApiKeyStatusHint.tsx
@@ -1,5 +1,5 @@
-import type { CanvasModelProviderStatus } from '../../types';
-import { useI18n } from '../../i18n';
+import type { CanvasModelProviderStatus } from '../../../types';
+import { useI18n } from '../../../i18n';
 
 interface ApiKeyStatusHintProps {
   status?: CanvasModelProviderStatus;
