+++ apps/canvas-workspace/src/renderer/src/components/settings/Settings/index.tsx
@@ -16,10 +16,10 @@
 
 import { useEffect, useState } from 'react';
 import { Drawer, SegmentedControl } from '../../ui';
-import { ModelsSection } from '../../chat/ModelSettings';
+import { ModelsSection } from '../ModelSettings';
 import { useCanvasModels } from '../../../hooks/useCanvasModels';
-import { ReplyStyleSection, usePromptProfile } from '../../chat/PromptSettings';
-import { RolesSection } from '../../chat/RolesSettings';
+import { ReplyStyleSection, usePromptProfile } from '../PromptSettings';
+import { RolesSection } from '../RolesSettings';
 import { AgentSection } from './AgentSection';
 import { BrowserSection } from './BrowserSection';
 import { BuiltInToolsSection } from './BuiltInToolsSection';
